refactor: styled component로 선언한 파일 네이밍 styled로 수정
