+++ src/components/cardPreview/CardPreview.styled.ts
