+++ src/components/newCardInputSection/NewCardInputSection.styled.ts
