+++ src/pages/newCardPage/NewCardPage.styled.ts
